dataprov_fs: make commands of the scdc_dataset_fs_store class usable when using fs_store provider
Looks like a copy-paste error from scdc_dataprov_fs_access::dataset_open().

Prior to this commit, any store opened with the fs_store data provider used the
commands as implemented in the scdc_dataprov_fs_access class, which, however,
are intended for the fs_access provider (fs aliases to this one, this is why the
filesystem test still works).

Spotted during development of the benchmark.

Signed-off-by: Thomas Weber <pokehaxco@gmail.com>